+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/PolarisContextAutoConfiguration.java
@@ -47,4 +47,8 @@ public class PolarisContextAutoConfiguration {
 		return new ModifyAddress();
 	}
 
+	@Bean
+	public ServiceRuleManager serviceRuleManager(SDKContext sdkContext) {
+		return new ServiceRuleManager(sdkContext);
+	}
 }