+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Change Log
 ---
 
-- [fix:fix wrong isAliveFlag config when creating new PolarisServer.](https://github.com/Tencent/spring-cloud-tencent/pull/179)
+- [Feature: Support parse ratelimit rule expression labels.](https://github.com/Tencent/spring-cloud-tencent/pull/182)